+++ backend/src/test/java/com/happy/friendogly/club/service/ClubQueryServiceTest.java
@@ -25,7 +25,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-
 class ClubQueryServiceTest extends ClubServiceTest {
 
     @Autowired
@@ -383,7 +382,8 @@ class ClubQueryServiceTest extends ClubServiceTest {
                 () -> assertThat(response.memberDetails().get(0).id()).isEqualTo(savedMember.getId()),
                 () -> assertThat(response.petDetails().size()).isEqualTo(1),
                 () -> assertThat(response.petDetails().get(0).id()).isEqualTo(savedPet.getId()),
-                () -> assertThat(response.allowedGender()).containsExactlyInAnyOrder(Gender.FEMALE, Gender.FEMALE_NEUTERED),
+                () -> assertThat(response.allowedGender()).containsExactlyInAnyOrder(Gender.FEMALE,
+                        Gender.FEMALE_NEUTERED),
                 () -> assertThat(response.allowedSize()).containsExactlyInAnyOrder(SizeType.SMALL)
         );
     }
